// constants/mtg/mtg-set-types.ts
export type MTGSetType =
  | "expansion"
  | "masters"
  | "promo"
  | "memorabilia"
  | "token"
  | "core"
  | "commander"
  | "draft_innovation"
  | "alchemy"
  | "minigame"
  | "masterpiece"
  | "funny"
  | "box"
  | "arsenal"
  | "starter"
  | "spellbook"
  | "planechase"
  | "duel_deck"
  | "from_the_vault"
  | "archenemy"
  | "treasure_chest"
  | "premium_deck"
  | "vanguard";

export enum MTGSetTypes {
  EXPANSION = "expansion",
  MASTERS = "masters",
  PROMO = "promo",
  MEMORABILIA = "memorabilia",
  TOKEN = "token",
  CORE = "core",
  COMMANDER = "commander",
  DRAFT_INNOVATION = "draft_innovation",
  ALCHEMY = "alchemy",
  MINIGAME = "minigame",
  MASTERPIECE = "masterpiece",
  FUNNY = "funny",
  BOX = "box",
  ARSENAL = "arsenal",
  STARTER = "starter",
  SPELLBOOK = "spellbook",
  PLANECHASE = "planechase",
  DUEL_DECK = "duel_deck",
  FROM_THE_VAULT = "from_the_vault",
  ARCHENEMY = "archenemy",
  TREASURE_CHEST = "treasure_chest",
  PREMIUM_DECK = "premium_deck",
  VANGUARD = "vanguard",
}